# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: TabMail API
  version: "2.0"
  description: |
    TabMail is a multi-tenant, API-first disposable email service with both
    inbound (SMTP receive) and outbound (SMTP send) capabilities.

    ## Quick Start for Bot/Automation Users

    ### Step 1: Get an API Key
    ```bash
    # Login to get a JWT token
    curl -X POST https://your-tabmail/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email":"you@example.com","password":"your-password"}'

    # Create an API Key with the JWT
    curl -X POST https://your-tabmail/api/v1/keys \
      -H "Authorization: Bearer <jwt_token>" \
      -H "Content-Type: application/json" \
      -d '{"label":"my-bot","scopes":["mailboxes:write","messages:read","send:write"]}'
    ```

    ### Step 2: Create a Disposable Mailbox
    ```bash
    # Get a suggested address
    curl https://your-tabmail/api/v1/domains/{domain_id}/suggest-address \
      -H "X-API-Key: tabmail_..."

    # Create the mailbox
    curl -X POST https://your-tabmail/api/v1/mailboxes \
      -H "X-API-Key: tabmail_..." \
      -H "Content-Type: application/json" \
      -d '{"full_address":"suggested-addr@domain.com","zone_id":"..."}'
    ```

    ### Step 3: Receive Emails (Poll or SSE)
    ```bash
    # Poll for messages
    curl https://your-tabmail/api/v1/mailbox/addr@domain.com \
      -H "X-API-Key: tabmail_..."

    # Or use Server-Sent Events for real-time
...
      -H "X-API-Key: tabmail_..."
    ```

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - APIKey: []
components:
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key
    MailboxToken:
      type: http
      scheme: bearer
      bearerFormat: TabMail-Mailbox-Token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant API key (alias for APIKey)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - QUOTA_EXCEEDED
                - DOMAIN_NOT_VERIFIED
                - SEND_NOT_ALLOWED
                - INTERNAL
            message: { type: string }
    Meta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
...
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        mailbox_id: { type: string, format: uuid }
        sender: { type: string }
        recipients: { type: array, items: { type: string } }
        subject: { type: string }
        size: { type: integer }
        seen: { type: boolean }
        headers: { type: object }
        received_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
    MessageDetail:
      allOf:
        - $ref: "#/components/schemas/Message"
        - type: object
          properties:
            text_body: { type: string }
            html_body: { type: string }
...
    OutboundState:
      type: string
      enum: [pending, processing, sent, retry, failed, dead]
    OutboundJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        mail_from:
          type: string
        rcpt_to:
          type: array
          items:
            type: string
        subject:
          type: string
        state:
          $ref: '#/components/schemas/OutboundState'
        attempts:
          type: integer
        max_attempts:
          type: integer
        last_error:
          type: string
        smtp_code:
          type: integer
        smtp_response:
          type: string
        message_id_header:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Permission Schemas ───────────────────────────────
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 30
        minimum: 1
        maximum: 100
      description: Items per page
paths:
...
  /api/v1/mailbox/{address}:
    get:
      tags: [Messages]
      summary: List messages in a mailbox
      operationId: listMessages
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
          example: "user@mail.example.com"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
                  meta: { $ref: "#/components/schemas/Meta" }
    delete:
      tags: [Messages]
      summary: Purge all messages from a mailbox
      operationId: purgeMailbox
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Purged }
...
  /api/v1/mailbox/{address}/{id}:
    get:
      tags: [Messages]
      summary: Get message detail
      operationId: getMessage
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/MessageDetail" }
    patch:
      tags: [Messages]
      summary: Mark message as seen
      operationId: markSeen
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      seen: { type: boolean }
    delete:
      tags: [Messages]
      summary: Delete a single message
      operationId: deleteMessage
      security: [ {}, { APIKey: [] }, { BearerAuth: [] }, { MailboxToken: [] } ]
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
...
  /api/v1/outbound:
    get:
      tags: [Sending]
      summary: List outbound jobs
      operationId: listOutboundJobs
      description: List outbound email delivery jobs for the current tenant.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Paginated list of outbound jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OutboundJob'
                  total:
                    type: integer
  /api/v1/outbound/{id}:
    get:
      tags: [Sending]
      summary: Get outbound job status
      operationId: getOutboundJob
      description: Get the delivery status of a specific outbound email job.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Outbound job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboundJob'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Admin: Domain Resources ─────────────────────────
...